Better tooltip caching

+++ backend/backend/api/auto_rois.py
@@ -15,6 +15,7 @@ from skimage.feature import blob_dog # blob_log, blob_doh
 
 from requests.utils import unquote
 from flask import request, jsonify
+from PIL import Image
 
 from qaboard.api import url_to_dir 
 from backend import app
@@ -26,7 +27,12 @@ def cached_read_image(image_path):
   """
   Simple LRU cache - the downside is that our images are huge so with 8 workers each saving 2 image, each 300MB, it's bad...
   """
-  image, meta = read_image(image_path)
+  image_pil = Image.open(str(file))
+  image = np.array(image_pil)
+  meta = {
+    # https://pillow.readthedocs.io/en/5.1.x/handbook/concepts.html#concept-modes
+    "mode": image_pil.mode
+  }
   return image, meta
 
 
@@ -50,12 +56,14 @@ image_cache_dir.mkdir(exist_ok=True, parents=True)
 
 def clear_memmapped_cache_dir():
     # keep only last 32 images
-    file_infos = list(image_cache_dir.glob('*.json'))
-    file_infos.sort(key=lambda f: -f.stat().st_mtime)
-    print(file_infos)
-    for file_info in file_infos[32:]:
-        file_info.unlink()
-        file_info.with_suffix('.dat').unlink()
+    file_data = list(image_cache_dir.glob('*.dat'))
+    file_data.sort(key=lambda f: -f.stat().st_mtime) # oldest last
+    for file in file_data[24:]:
+        print(f"RM {file}")
+        file.unlink()
+        file_info = file.with_suffix('.json')
+        if file_info.exists():
+          file_info.unlink()
 
 def memmapped_read_image(image_path):
   key = f"{image_path}-{image_path.stat().st_mtime}"
@@ -63,32 +71,32 @@ def memmapped_read_image(image_path):
   image_cache_data = image_cache_dir / f"{hash}.dat"
   image_cache_info = image_cache_dir / f"{hash}.json"
   if not (image_cache_data.exists() and image_cache_info.exists()):
-    if under_uwsgi:
-      # worst case the 1st requests will write multiple times that file...
-      uwsgi.lock()
+    clear_memmapped_cache_dir()
+    # if under_uwsgi:
+    #   # worst case the 1st requests will write multiple times that file...
+    #   uwsgi.lock()
+    # print(f'MISS {image_path}')
     image, meta = read_image(image_path)
+    # print(f'READ')
     with image_cache_info.open('w') as fmeta:
       json.dump({"meta": meta, "shape": image.shape, "dtype": str(image.dtype)}, fmeta)
-    fp = np.memmap(image_cache_data, dtype=image.dtype, mode='w+', shape=image.shape)
+    fp = np.memmap(image_c  ache_data, dtype=image.dtype, mode='w+', shape=image.shape)
     fp[:] = image[:]
     fp.flush() # write to disk
-    if under_uwsgi:
-      uwsgi.unlock()
+    # print(f'WRITE')
+    # if under_uwsgi:
+    #   uwsgi.unlock()
     return fp, meta
   else:
-    if hash.endswith('0'): # 1/16 chance... is it too often?
-      clear_memmapped_cache_dir()
-
+    # print(f'HIT {hash}')
     with image_cache_info.open() as f:
       info = json.load(f)
     fp = np.memmap(image_cache_data, dtype=info['dtype'], mode='r', shape=tuple(info['shape']))
     return fp, info['meta']
 
 
-
 @app.route("/api/v1/output/image/pixel", methods=['GET', 'POST'])
 def get_pixel():
-  
   x = int(request.args['x'])-1
   y = int(request.args['y'])-1
   image_path = url_to_dir(request.args['image_url'])
